modal de alerta de arreglado
