+++ src/components/Alumnes/BotonConfirmarInscriptes.jsx
@@ -1,163 +1,36 @@
 import React from 'react'
-import { Button, Modal} from 'semantic-ui-react';
-import axios from 'axios';
-const URL_Inscriptos = `${process.env.REACT_APP_API_URL}/inscriptos/`;
+import { Button, Confirm  } from 'semantic-ui-react';
+
 class BotonConfirmarInscriptes extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      confirmacion:this.props.callback,
-      active: true,
-      inscriptes:this.props.inscriptes,
-      curso:this.props.cursoActual
-    }
-    console.log(this.props.cursoActual);
-  }
+  state = { open: false}
 
-  onCloseModal(estado) {
-    this.setState({
-      open: estado
-    });
+  show = () => this.setState({ open: true })
+  handleConfirm = () =>{
+    this.props.onRegister("confirmed");
+    this.setState({open: false });
   }
-  onSubmit = async (onRegistrarCorrectamente) =>{
-    let curso = this.props.cursoActual;
-    let lista = this.state.inscriptes;
-    let listaNueva = [];
-    console.log(lista);
-
-    const API_URL = `${process.env.REACT_APP_API_URL}/cursos/${curso}/inscriptes`;
-    await
-    axios
-      .get(`${API_URL}`)
-      .then(response => {
-        listaNueva = response.data.response;
-        listaNueva.forEach(inscripte => {
-          axios
-          .delete(`${process.env.REACT_APP_API_URL}/estudiantes/${inscripte.estudiante.id}?curseId=${curso}`)
-       })
-      })
-      .catch(function (error) {
-        console.log(error);
-      });
-    lista.forEach(inscripte => {
-      var nuevoInscripte = {
-        "estudianteId": inscripte.id,
-        "cursoId": curso
-      }
-      fetch(`${URL_Inscriptos}`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'Content-Length': JSON.stringify(nuevoInscripte).length.toString()
-        },
-        body: JSON.stringify(nuevoInscripte)
-      })
-    });
-    this.onCloseModal(false)
-    // window.location.href = window.location.href;
-  }
-
+  handleCancel = () =>{
+    this.props.onRegister("cancelled");
+    this.setState({open: false });
+  } 
   render() {
+    const { open } = this.state
+
     return (
       <div>
-        <Modal
-          onClose={() => this.onCloseModal(false)}
-          onOpen={() => this.onCloseModal(true)}
-          open={this.state.open}
-          trigger={<Button className="confirmButton">Confirmar</Button>}
-        >
-          <Modal.Header>Alerta!</Modal.Header>
-          <Modal.Content>
-            <Modal.Description>
-              <p>
-                Esta seguro que desea subir un nuevo csv, se eliminara los alumnes incritos actuales.
-			  </p>
-            </Modal.Description>
-          </Modal.Content>
-          <Modal.Actions>
-            <Button className="cancelButton" onClick={() => this.onCloseModal(false)}>
-              Cancelar
-			</Button>
-            <Button className="confirmButton"
-              content="Confimar"
-              labelPosition='right'
-              onClick={() => this.onSubmit()}
-              positive
-            />
-          </Modal.Actions>
-        </Modal>
+        <Button className="confirmButton" onClick={this.show}>Confirmar</Button>
+        <Confirm
+          content={"¿Esta seguro que desea importar un nuevo csv?, se eliminaran los alumnes previamente inscritos."}
+          open={open}
+          header={"Alerta!"}
+          cancelButton={"Cancelar"}
+          confirmButton={"Confirmar"}
+          onCancel={this.handleCancel}
+          onConfirm={this.handleConfirm}
+        />
       </div>
-    );
+    )
   }
 }
-
-
-// const URL_Inscriptos = `${process.env.REACT_APP_API_URL}/inscriptos/`;
-// async function onSubmit(onRegistrarCorrectamente){
-//     // let curso = 1;
-//     // let lista = this.state.inscriptes;
-//     // let listaNueva = [];
-//     // const API_URL = `${process.env.REACT_APP_API_URL}/cursos/${curso}/inscriptes`;
-//     // await
-//     // axios
-//     //   .get(`${API_URL}`)
-//     //   .then(response => {
-//     //     listaNueva = response.data.response;
-//     //     listaNueva.forEach(inscripte => {
-//     //       axios
-//     //       .delete(`${process.env.REACT_APP_API_URL}/estudiantes/${inscripte.estudiante.id}?curseId=${curso}`)
-//     //    })
-//     //   })
-//     //   .catch(function (error) {
-//     //     console.log(error);
-//     //   });
-//     // lista.forEach(inscripte => {
-//     //   var nuevoInscripte = {
-//     //     "estudianteId": inscripte.id,
-//     //     "cursoId": curso
-//     //   }
-//     //   fetch(`${URL_Inscriptos}`, {
-//     //     method: 'POST',
-//     //     headers: {
-//     //       'Content-Type': 'application/json',
-//     //       'Content-Length': JSON.stringify(nuevoInscripte).length.toString()
-//     //     },
-//     //     body: JSON.stringify(nuevoInscripte)
-//     //   })
-//     // });
-//     // this.setOpen(false);
-// }
-
-// function ModalExampleShorthand() {
-// 	const [open, setOpen] = React.useState(false)
-// 	return (
-// 		<Modal
-// 		  onClose={() => setOpen(false)}
-// 		  onOpen={() => setOpen(true)}
-// 		  open={open}
-// 		  trigger={<Button className="confirmButton">Confirmar</Button>}
-// 		>
-// 		  <Modal.Header>Alerta!</Modal.Header>
-// 		  <Modal.Content>
-// 			<Modal.Description>
-// 			  <p>
-// 				Esta seguro que desea subir un nuevo csv, se eliminara los alumnes incritos actuales.
-// 			  </p>
-// 			</Modal.Description>
-// 		  </Modal.Content>
-// 		  <Modal.Actions>
-// 			<Button className="cancelButton" onClick={() => setOpen(false)}>
-// 			  Cancelar
-// 			</Button>
-// 			<Button className="confirmButton"
-// 			  content="Confimar"
-// 			  labelPosition='right'
-// 			  onClick= {onSubmit()}
-// 			  positive
-// 			/>
-// 		  </Modal.Actions>
-// 		</Modal>
-// 	  )
-// 	}
 
 export default BotonConfirmarInscriptes
