+++ src/components/Alumnes/BotonImportar.jsx
@@ -189,40 +189,40 @@ class BotonImportar extends Component {
     this.abrirModal(false)
   }
 
-  // onSubmit = async (onRegistrarCorrectamente) =>{
-  //   let curso = this.props.cursoActual;
-  //   let lista = this.state.inscriptes;
-  //   let listaNueva = [];
-  //   const API_URL = `${process.env.REACT_APP_API_URL}/cursos/${curso}/inscriptes`;
-  //   await
-  //   axios
-  //     .get(`${API_URL}`)
-  //     .then(response => {
-  //       listaNueva = response.data.response;
-  //       listaNueva.forEach(inscripte => {
-  //         axios
-  //         .delete(`${process.env.REACT_APP_API_URL}/estudiantes/${inscripte.estudiante.id}?curseId=${curso}`)
-  //      })
-  //     })
-  //     .catch(function (error) {
-  //       console.log(error);
-  //     });
-  //   lista.forEach(inscripte => {
-  //     var nuevoInscripte = {
-  //       "estudianteId": inscripte.id,
-  //       "cursoId": curso
-  //     }
-  //     fetch(`${URL_Inscriptos}`, {
-  //       method: 'POST',
-  //       headers: {
-  //         'Content-Type': 'application/json',
-  //         'Content-Length': JSON.stringify(nuevoInscripte).length.toString()
-  //       },
-  //       body: JSON.stringify(nuevoInscripte)
-  //     })
-  //   });
-  //   this.abrirModal(false)
-  // }
+  onSubmit = async (onRegistrarCorrectamente) =>{
+    let curso = this.props.cursoActual;
+    let lista = this.state.inscriptes;
+    let listaNueva = [];
+    const API_URL = `${process.env.REACT_APP_API_URL}/cursos/${curso}/inscriptes`;
+    await
+    axios
+      .get(`${API_URL}`)
+      .then(response => {
+        listaNueva = response.data.response;
+        listaNueva.forEach(inscripte => {
+          axios
+          .delete(`${process.env.REACT_APP_API_URL}/estudiantes/${inscripte.estudiante.id}?curseId=${curso}`)
+       })
+      })
+      .catch(function (error) {
+        console.log(error);
+      });
+    lista.forEach(inscripte => {
+      var nuevoInscripte = {
+        "estudianteId": inscripte.id,
+        "cursoId": curso
+      }
+      fetch(`${URL_Inscriptos}`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'Content-Length': JSON.stringify(nuevoInscripte).length.toString()
+        },
+        body: JSON.stringify(nuevoInscripte)
+      })
+    });
+    this.abrirModal(false);
+  }
 
 
   handleOnDrop = (data) => {
@@ -283,7 +283,15 @@ class BotonImportar extends Component {
   getResponse(result) {
     this.abrirModal(false)
     console.log(result);
+}
+register=(result)=>
+{
+  if(result==="confirmed")
+  {
+   this.onSubmit();
+   
   }
+}
 
   render() {
     return (
@@ -332,8 +340,7 @@ class BotonImportar extends Component {
         }
         <Modal.Actions>
           <Button className="cancelButton" onClick={() => this.abrirModal(false)}>Cerrar</Button>
-          {/* <Button className="confirmButton" onClick={() => this.onSubmit()}>Confirmar</Button> */}
-          <BotonConfirmarInscriptes inscriptes={this.state.inscriptes} callback={this.getResponse.bind(this)} cursoActual={this.state.curso} />
+          <BotonConfirmarInscriptes onRegister={this.register}></BotonConfirmarInscriptes>
         </Modal.Actions>
       </Modal>
     )
